# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ТехноСпортФест 2026 API",
    "version": "1.0.0",
    "description": "Канонический контракт MVP. Времена UTC. JSON camelCase. Неизвестные поля запросов отклоняются. Списки сортируются детерминированно: competitions по startsAt,id; participants по fullName,registrationId; results по place,fullName,id; news/documents по publishedAt DESC,id DESC; rating по points DESC,fullName,id."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/admin/permissions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Каталог прав",
        "description": "Группированный справочник прав для конструктора роли. Только roles.manage.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionGroup"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "operationId": "listPermissions"
      }
    },
    "/admin/roles": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Список ролей",
        "description": "Все роли с их правами. Только roles.manage.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "operationId": "listRoles"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Создать роль",
        "description": "Создаёт динамическую роль с набором прав. Только roles.manage. Системные роли менять нельзя.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-origin-check": true,
        "operationId": "createRole",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          }
        }
      }
    },
    "/admin/roles/{roleId}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Изменить роль",
        "description": "Меняет имя/описание динамической роли. Только roles.manage; системные роли — 403 ROLE_SYSTEM.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-origin-check": true,
        "operationId": "updateRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePatch"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Удалить роль",
        "description": "Только динамические роли без назначенных пользователей. Системная — 403 ROLE_SYSTEM, назначенная — 409 ROLE_IN_USE.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "Успех"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        },
        "x-origin-check": true,
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/admin/roles/{roleId}/permissions": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Назначить права роли",
        "description": "Полностью заменяет набор прав динамической роли. Только roles.manage; системные роли — 403 ROLE_SYSTEM.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-origin-check": true,
        "operationId": "updateRolePermissions",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionsUpdate"
              }
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Список пользователей",
        "description": "Пользователи с ролью и именем (из профиля спортсмена, если есть). Только users.manage.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "operationId": "listAdminUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ]
      }
    },
    "/admin/users/{userId}/role": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Назначить роль пользователю",
        "description": "Меняет роль пользователя. Изменение прав вступит в силу после следующего входа (JWT до 24ч). Только users.manage.",
        "x-owner": "B1",
        "x-milestone": "v5",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "x-origin-check": true,
        "operationId": "assignUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tsf_access",
        "description": "JWT, HttpOnly; SameSite=Lax; Secure в production; срок 24 часа"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable Russian message"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional field errors"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "PermissionGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "section": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "section",
          "label",
          "codes"
        ]
      },
      "Role": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "isSystem",
          "permissions",
          "createdAt"
        ]
      },
      "RoleCreate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "RolePatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          }
        }
      },
      "PermissionsUpdate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissions"
        ]
      },
      "AssignRoleRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "roleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "roleId"
        ]
      },
      "AdminUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "roleName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "roleId",
          "roleName",
          "createdAt"
        ]
      },
      "AdminUserPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      }
    },
    "responses": {
...
      "401": {
        "description": "Нет действующего JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "UNAUTHENTICATED",
              "message": "Нет действующего JWT"
            }
          }
        }
      },
      "403": {
        "description": "Недостаточно прав или неверный/отсутствующий Origin",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "FORBIDDEN",
              "message": "Недостаточно прав или неверный/отсутствующий Origin"
            }
          }
        }
      },
      "404": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "NOT_FOUND",
              "message": "Ресурс не найден"
            }
          }
        }
      },
      "409": {
        "description": "Конфликт состояния или уникальности",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "CONFLICT",
              "message": "Конфликт состояния или уникальности"
            }
          }
        }
      },
      "422": {
        "description": "Ошибка полей",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "VALIDATION_ERROR",
              "message": "Ошибка полей"
            }
          }
        }
      }
    }
  }
}